+++ printing/lib/src/preview/controller.dart
@@ -67,7 +67,9 @@ class PdfPreviewData extends ChangeNotifier {
   PdfPageFormat get actualPageFormat => _onComputeActualPageFormat();
 
   String get localPageFormat {
-    final locale = WidgetsBinding.instance!.window.locale;
+    // Flutter 3 workaround
+    final WidgetsBinding? wbi = WidgetsBinding.instance;
+    final locale = wbi!.window.locale;
     // ignore: unnecessary_cast
     final cc = (locale as Locale?)?.countryCode ?? 'US';